requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.11
django-cors-headers==2.5.2
django-crispy-forms==1.7.2
django-filter==1.0.4
django-formtools==2.1
django-guardian==1.5.0
django-redis==4.10.0
django-reversion==3.0.3
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
drf-extensions==0.4.0
future==0.17.1
httplib2==0.12.1
idna==2.8
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.8
MarkupSafe==1.1.1
Pillow==5.4.1
pycryptodome==3.8.0
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2018.9
raven==6.10.0
redis==3.2.1
requests==2.21.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
XlsxWriter==1.1.5
xlwt==1.3.0
simplejson==3.16.0